Extract flexible space
