+++ packages/dx-react-grid-bootstrap3/src/templates/toolbar/flexible-space.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const FlexibleSpace = () => <div style={{ flex: '0 0 0' }} />;
