+++ packages/dx-react-grid-bootstrap3/src/templates/toolbar/toolbar.jsx
